+++ packages/android_alarm_manager/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Remove support for the V1 Android embedding.
+* Updated Android lint settings.
 
 ## 2.0.2